[CHORE]: 파일명 수정 + 그에 따른 코드 수정
