+++ ARtique/ARtique/Networtk/MoyaTarget/AuthService.swift
@@ -1,5 +1,5 @@
 //
-//  LoginService.swift
+//  AuthService.swift
 //  ARtique
 //
 //  Created by hwangJi on 2022/04/27.
@@ -8,11 +8,12 @@
 import Foundation
 import Moya
 
-enum LoginService {
+enum AuthService {
     case kakaoLogin(refreshToken: String)
+    case renewalToken(refreshToken: String)
 }
 
-extension LoginService: TargetType {
+extension AuthService: TargetType {
     var baseURL: URL {
         return URL(string: APIConstants.baseURL)!
     }
@@ -21,19 +22,21 @@ extension LoginService: TargetType {
         switch self {
         case .kakaoLogin:
             return "/auth/kakaoLogin"
+        case .renewalToken:
+            return "/auth/token"
         }
     }
     
     var method: Moya.Method {
         switch self {
-        case .kakaoLogin:
+        case .kakaoLogin, .renewalToken:
             return .post
         }
     }
     
     var task: Task {
         switch self {
-        case .kakaoLogin(let refreshToken):
+        case .kakaoLogin(let refreshToken), .renewalToken(let refreshToken):
             return .requestParameters(parameters: ["refreshToken": refreshToken], encoding: JSONEncoding.default)
         }
     }
